Add monster search screen with autocomplete
